chore: move wallet connect context back to provider file
